requirements.txt
anthropic>=0.40.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
chainlit>=1.3.0
python-dotenv>=1.0.0
pypdf>=4.0.0
python-docx>=1.0.0
